fix rcd and error handling

+++ gaskuyy-quarkus/src/main/java/com/abdun/SrvCart.java
@@ -2,12 +2,17 @@ package com.abdun;
 
 import java.util.List;
 
+import javax.management.RuntimeErrorException;
+
+import io.vertx.mutiny.ext.web.handler.ErrorHandler;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.NoResultException;
 import jakarta.persistence.TypedQuery;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
 
 /**
  *
@@ -21,7 +26,7 @@ public class SrvCart {
 	private EntityManager em;
 
 	public void update(RcdCart rcdCart) {
-		em.merge(rcdCart);
+			em.merge(rcdCart);
 	}
 
 	public void detach(Object record) {
@@ -29,25 +34,42 @@ public class SrvCart {
 	}
 
 	public List<RcdCart> getProductsFromCart() {
-		try {
-			TypedQuery<RcdCart> tq = em.createQuery("SELECT h FROM RcdCart h ORDER BY h.id DESC",
+			try {
+				TypedQuery<RcdCart> tq = em.createQuery("SELECT h FROM RcdCart h ORDER BY h.id DESC",
 					RcdCart.class);
 			return tq.getResultList();
-		} catch (NoResultException e) {
-			return null;
-		}
+			} catch (Exception e) {
+				return null;
+			}
 	}
 
 	public RcdCart findById(int id) {
-		RcdCart rcdCart = em.find(RcdCart.class, id);
-		return rcdCart;
+		try {
+			RcdCart rcdCart = em.find(RcdCart.class, id);
+			return rcdCart;
+		} catch (Exception e) {
+			return null;
+		}
+	}
+	
+	public RcdCart findByProductId(int id) {
+		try {
+			TypedQuery<RcdCart> tq = em.createQuery("SELECT h FROM RcdCart h WHERE h.productId = :id",
+					RcdCart.class);	
+					tq.setParameter("id", id);
+			return tq.getSingleResult();
+		} catch (Exception e) {
+			return null;
+		}
+		
 	}
 
 	public void delete(int id) {
-		System.out.println("remove cart");
-		RcdCart r = findById(id);
-		// detach(r);
-		System.out.println("data " + r.getId());
-		em.remove(r);
+			RcdCart r = findById(id);
+			if (r != null) {
+				em.remove(r);	
+			} else {
+				throw new WebApplicationException("Gagal hapus data produk",Response.Status.NOT_FOUND); 
+			}
 	}
 }
